refactor: split media delegates into extensions
