+++ iOSClient/Media/NCMedia.swift
@@ -508,58 +508,3 @@ class NCMedia: UIViewController {
         print("[MEDIA CACHE] END center: \(centerIndex) hits: \(cacheHits) diskReads: \(diskReads) knownMissing: \(knownMissingImages) newMissing: \(newMissingImages) loaded: \(loadedImages)")
     }
 }
-
-// MARK: -
-
-extension NCMedia: UIScrollViewDelegate {
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        setTitleDate()
-
-        if !dataSource.compactMetadatas.isEmpty {
-            setNeedsStatusBarAppearanceUpdate()
-        }
-    }
-
-    func scrollViewDidEndDragging(
-        _ scrollView: UIScrollView,
-        willDecelerate decelerate: Bool
-    ) {
-        if !decelerate {
-            updateImageCacheWindow()
-            searchNewMedia()
-        }
-    }
-
-    func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
-        updateImageCacheWindow()
-        searchNewMedia()
-    }
-
-    func scrollViewDidEndScrollingAnimation(_ scrollView: UIScrollView) {
-        lastCacheCenterIndex = nil
-        updateImageCacheWindow()
-    }
-}
-
-// MARK: -
-
-extension NCMedia: NCSelectDelegate {
-    func dismissSelect(serverUrl: String?, metadata: tableMetadata?, type: String, items: [Any], overwrite: Bool, copy: Bool, move: Bool, session: NCSession.Session, controller: NCMainTabBarController?) {
-        guard let serverUrl else { return }
-
-        Task {
-            let home = utilityFileSystem.getHomeServer(session: session)
-            let mediaPath = serverUrl.replacingOccurrences(of: home, with: "")
-
-            await database.setAccountMediaPathAsync(mediaPath, account: session.account)
-
-            self.imageCache.removeAll()
-            self.missingImageCacheKeys.removeAll()
-
-            await self.debouncerLoadDataSource.call {
-                await self.loadDataSource()
-            }
-            searchNewMedia()
-        }
-    }
-}
